% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{clean_names}
\alias{clean_names}
\title{Clean column names
Used in func:create_mapping}
\usage{
clean_names(oldName)
}
\arguments{
\item{oldName}{a string}
}
\value{
a string
}
\description{
Clean column names
Used in func:create_mapping
}
