% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram_func.R
\name{codeGenerate}
\alias{codeGenerate}
\title{Generate a Label Code}
\usage{
codeGenerate(df)
}
\arguments{
\item{df}{a dataframe containing a column "parent"}

\item{symbol}{a character string used in the instance code (e.g. 'E' for
events)}
}
\value{
data.frame containing a column "code"
}
\description{
`codeGenerate()` is a function used to generate a code for each instances
used in a graph or subgraph. This code is used to simplify the template
syntax.
}
\concept{string processing function}
\keyword{string}
