% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{concat}
\alias{concat}
\title{Replace Fieldname
Used in func: rename_df, make_temporal, createLongFileEvent,
createLongFileOcc}
\usage{
concat(df, col1, col2, sep)
}
\arguments{
\item{name}{a string (fieldname)}

\item{mapping}{a data.frame (to map the fieldname with a standard name)}
}
\value{
a string (darwinized fieldname)
}
\description{
Replace Fieldname
Used in func: rename_df, make_temporal, createLongFileEvent,
createLongFileOcc
}
