% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{convert_columns}
\alias{convert_columns}
\title{Convert Column values
Used in module: mod_Import}
\usage{
convert_columns(df, userDic)
}
\arguments{
\item{df}{a data.frame}

\item{userDic}{a data.frame, used to convert old values in new values
in df}
}
\value{

}
\description{
Convert Column values
Used in module: mod_Import
}
