% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{create_df_level}
\alias{create_df_level}
\title{Create Levels Relationships
Used in func: find_parent}
\usage{
create_df_level(df)
}
\arguments{
\item{df}{a data.frame, output of createAggregatedDF()}
}
\value{
a data.frame (columns: parent, children)
}
\description{
Create Levels Relationships
Used in func: find_parent
}
