% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{create_group_cols}
\alias{create_group_cols}
\title{Create list of Event/Occ fieldnames}
\usage{
create_group_cols(dtf.event, dtf.occurrence)
}
\arguments{
\item{dtf.event}{a list (Event parameters)}

\item{dtf.occurrence}{a list (Occurrence parameters)}
}
\value{
a list of vectors, containing Event and Occurrence fieldnames
}
\description{
Create list of Event/Occ fieldnames
}
