% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{create_mapping}
\alias{create_mapping}
\title{Map fieldnames to Darwin core names
`create_mapping()` is a function create the dictionnary to map columns names
from the user dataset to standard from Darwin Core
Used in module: mod_Import}
\usage{
create_mapping(dico_init, list_user)
}
\arguments{
\item{dico_init}{a data.frame containing standard names}

\item{list_user}{a vector of strings, containing fieldnames}
}
\value{

}
\description{
Map fieldnames to Darwin core names
`create_mapping()` is a function create the dictionnary to map columns names
from the user dataset to standard from Darwin Core
Used in module: mod_Import
}
