% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{createAggregatedDF_v1}
\alias{createAggregatedDF_v1}
\title{Create Aggregated Data.Frame}
\usage{
createAggregatedDF_v1(df, dtf.event, dtf.occurrence)
}
\arguments{
\item{df}{a data.frame}

\item{dtf.event}{a list (Event parameters)}

\item{dtf.occurrence}{a list (Occurrence parameters)}
}
\value{
a data.frame
}
\description{
Create Aggregated Data.Frame
}
