% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram_func.R
\name{createGraph}
\alias{createGraph}
\title{Create a Graph}
\usage{
createGraph(dfRelations, dfNodeLabel)
}
\arguments{
\item{dfRelations}{a dataframe containing heterogeneous relations (between
two different groups of records, that contains at least one column "parent"
and one column "children").}

\item{dfNodeLabel}{a dataframe that contains the labels of instances (which
will be used for visual representation) and a code}
}
\value{
a character string containing the heterogeneous relations in the
form parent -> children, to insert into a template for grViz() function.
}
\description{
`createGraph()` is a function used to transform create subgraph based on
DiagrammeR syntax and template. It is used to create one subgraph per group
of records according to the Darwin Core scheme (i.e. Event Core, Occurrence
and eMoF extensions). It models homogeneous relations (several levels in the
same group of records). It returns a character string to insert into
a template readable by grViz() function.
}
\seealso{
Other DiagrammeR function: 
\code{\link{createSubgraph}()}
}
\concept{DiagrammeR function}
\keyword{graph}
