% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{createLongFileEvent}
\alias{createLongFileEvent}
\title{Create the Event Data.Frame}
\usage{
createLongFileEvent(df_agg, dtf.event, mapping)
}
\arguments{
\item{df_agg}{data.frame, output of createAggregatedDF()}

\item{dtf.event}{a list of Event parameters}

\item{mapping}{a data.frame (darwinizer mapping)}
}
\value{
a data.frame
}
\description{
`createLongFileEvent()`is a function used to create the Event data.frame
(will be exported).
}
\seealso{
Other organisation function: 
\code{\link{createLongFileOcc}()}
}
\concept{organisation function}
