% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{createLongFileOcc}
\alias{createLongFileOcc}
\title{Create the Occurrence Data.Frame}
\usage{
createLongFileOcc(df_agg, dtf.event, dtf.occurrence, mapping)
}
\arguments{
\item{df_agg}{data.frame, output of createAggregatedDF()}

\item{dtf.event}{a list of Event parameters}

\item{dtf.occurrence}{a list of Occurrence parameters}

\item{mapping}{a data.frame (darwinizer mapping)}
}
\value{
a data.frame
}
\description{
`createLongFileOcc()`is a function used to create the Occurrence data.frame
(will be exported).
}
\seealso{
Other organisation function: 
\code{\link{createLongFileEvent}()}
}
\concept{organisation function}
