% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{createMoF}
\alias{createMoF}
\title{Create the first part of Measurement data.frame (1st level MoF)}
\usage{
createMoF(dtf.info, df_agg, typeLevel = "Event")
}
\arguments{
\item{dtf.info}{list of Event or Occurrence parameters}

\item{df_agg}{data.frame, output of createAggregatedDF()}

\item{typeLevel}{a string, either "Event" or "Occurrence"}
}
\value{
a data.frame
}
\description{
Create the first part of Measurement data.frame (1st level MoF)
}
