% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{createMoFMoF}
\alias{createMoFMoF}
\title{Title}
\usage{
createMoFMoF(dtf.info, dfMoF, df_agg, typeLevel = "Event")
}
\arguments{
\item{dtf.info}{list of Event or Occurrence parameters}

\item{dfMoF}{data.frame, output of createMoF()}

\item{df_agg}{data.frame, output of createAggregatedDF()}

\item{typeLevel}{a string, either "Event" or "Occurrence"}
}
\value{
a data.frame
}
\description{
Title
}
