% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{createNameLevels}
\alias{createNameLevels}
\title{Create list of Levels
Used in func: createAggregatedDF}
\usage{
createNameLevels(dtf.event, dtf.occurrence)
}
\arguments{
\item{dtf.event}{a list (Event parameters)}

\item{dtf.occurrence}{a list (Occurrence parameters)}
}
\value{
a list of strings (either "Event" or "Occ")
}
\description{
Create list of Levels
Used in func: createAggregatedDF
}
