% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram_func.R
\name{createSubgraph}
\alias{createSubgraph}
\title{Create a Subgraph}
\usage{
createSubgraph(
  df,
  dfNodeLabel,
  level,
  labelSubgraph,
  numCluster,
  style = solid
)
}
\arguments{
\item{dfNodeLabel}{a dataframe that contains the labels of instances (which
will be used for visual representation) and a code}

\item{labelSubgraph}{a character string corresponding to the group of records
represented in the subgraph (i.e. Event Core, Occurrence ext. or eMoF ext.)}

\item{numCluster}{an integer corresponding to the subgraph number. The
subgraphs must be numbered from 0 to N in ascending order to avoid errors
when using grViz() function.}

\item{style}{a character string indicating the style of subgraph box border
(e.g. bold, solid, dashed)}

\item{dfRelations}{a dataframe specific to one group of record (homogeneous),
that contains at least one column "parent" and one column "children"}
}
\value{
a character string containing the subgraphs caracteristics, to insert
into a template for grViz() function.
}
\description{
`createSubgraph()` is a function used to transform create subgraph based on
DiagrammeR syntax and template. It is used to create one subgraph per group
of records according to the Darwin Core scheme (i.e. Event Core, Occurrence
and eMoF extensions). It models homogeneous relations (several levels in the
same group of records). It returns a character string to insert into
a template readable by grViz() function.
}
\seealso{
Other DiagrammeR function: 
\code{\link{createGraph}()}
}
\concept{DiagrammeR function}
\keyword{graph}
