% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram_func.R
\name{createSubgraphOld}
\alias{createSubgraphOld}
\title{Old versions}
\usage{
createSubgraphOld(
  dfRelations,
  dfNodeLabel,
  labelSubgraph,
  numCluster,
  style = solid
)
}
\description{
Old versions
}
