% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{dateIso}
\alias{dateIso}
\title{Transform Dates into ISO 8601 format, either datetime or interval}
\usage{
dateIso(d1, d2 = NA, simplify = TRUE, format_date = NA, sep_interval = NA)
}
\arguments{
\item{d1}{a character string corresponding to the beginning date of an
interval. Must be be "YYYY", "YYYY-MM", or "YYYY-MM-DD".}

\item{d2}{an optional character string corresponding to the ending date of an
interval. Must be be "YYYY", "YYYY-MM", or "YYYY-MM-DD".}

\item{simplify}{logical indicating if iso8601 format should be "YYYY-MM-DD"
even if the month and day are not provided (set to "01" by default)}

\item{format_date}{a character vector of date-time formats}

\item{sep_interval}{a character, indicating the interval between begin and end date}
}
\value{
a character object formatted according to ISO 8601
}
\description{
`dateIso()` is a function used to
Used in func: make_temporal
}
\concept{standardization function}
\keyword{date}
