% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{df2list}
\alias{df2list}
\title{df2list}
\usage{
df2list(df, title = NA)
}
\arguments{
\item{title}{name of the first item, closed. Optionnal, if it is not used, the first items of the tree will be shown}

\item{dtf}{a dataframe. Each column represent a hierarchical level of the tree to be built}
}
\description{
Transform a dataframe to the shape required by shinyTree().
}
