% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{duplicate_columns}
\alias{duplicate_columns}
\title{Duplicate a Column
Used in module: mod_Import}
\usage{
duplicate_columns(df, col_dupli)
}
\arguments{
\item{df}{a data.frame}

\item{col_dupli}{a list of strings}
}
\value{
a data.frame
}
\description{
Duplicate a Column
Used in module: mod_Import
}
