% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{exclude}
\alias{exclude}
\title{Exclude From A List}
\usage{
exclude(list, names)
}
\arguments{
\item{list}{a list}

\item{names}{a vector of strings}
}
\value{
a list
}
\description{
Exclude From A List
}
