% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{find_parent}
\alias{find_parent}
\title{Find the parent Level
Used in func: find_parent_event}
\usage{
find_parent(level, df)
}
\arguments{
\item{level}{a string, either "Occ" or "Event"}

\item{df}{df a data.frame, output of createAggregatedDF()}
}
\value{
a string, the name of the parent level (e.g. "Event-2")
}
\description{
Find the parent Level
Used in func: find_parent_event
}
