% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{make_temporal}
\alias{make_temporal}
\title{Create temporal data.frame}
\usage{
make_temporal(
  df,
  df_mapped,
  type,
  format_date = NA,
  sep_interval = NA,
  values = NA,
  tf_dic = NA
)
}
\arguments{
\item{df}{a data.frame (user)}

\item{df_mapped}{a data.frame (darwinizer mapping)}

\item{type}{a string (controlled list of values indicating the type of conversion)}

\item{format_date}{a character vector of date-time formats}

\item{sep_interval}{a character, indicating the interval between begin and end date}

\item{values}{list of two dates}

\item{tf_dic}{#not functional}
}
\value{

}
\description{
Create temporal data.frame
}
