% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{member}
\alias{member}
\title{List Member
Used in func: exclude}
\usage{
member(list, names)
}
\arguments{
\item{list}{a list}

\item{names}{a vector of strings}
}
\value{
a list
}
\description{
List Member
Used in func: exclude
}
