% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{orderDataFrame}
\alias{orderDataFrame}
\title{Order a Data.frame (row-based)
Used in func: createLongFileEvent, createLongFileOcc}
\usage{
orderDataFrame(df, col)
}
\arguments{
\item{df}{a data.frame}

\item{col}{a column name, basis to order data.frame}
}
\value{
a data.frame
}
\description{
Order a Data.frame (row-based)
Used in func: createLongFileEvent, createLongFileOcc
}
