% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{remove_col_all}
\alias{remove_col_all}
\title{Remove Empty Columns
Used in func: createLongFileEvent, createLongFileOcc}
\usage{
remove_col_all(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
a data.frame
}
\description{
Remove Empty Columns
Used in func: createLongFileEvent, createLongFileOcc
}
