% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{rename_base}
\alias{rename_base}
\title{Tidy Data.Frame}
\usage{
rename_base(df, old_colname, new_colname)
}
\arguments{
\item{df}{an untidy data.frame}

\item{untidy_col}{a vector of column names that are not tidy (e.g. "wild boars")}

\item{new_col}{a string indicating the new column names corresponding to the
untidy columns (e.g. "species")}

\item{new_value}{a string indicating the name of the value corresponding to the
untidy columns (e.g. "individualCount")}
}
\value{
a tidy data.frame of data, with new column names
}
\description{
`tidy_df()` is a function used to transform a data.frame into its tidy
version (one variable per column)
}
\details{
Used in func: tidy_data
}
\concept{pre-processing function}
