% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{rename_col_start}
\alias{rename_col_start}
\title{Rename Columns based on Prefix
Used in func: createMoF}
\usage{
rename_col_start(df, oldpref, newpref)
}
\arguments{
\item{df}{a data.frame to rename}

\item{oldpref}{a string, prefix to select column names}

\item{newpref}{a string, to replace selected column names}
}
\value{

}
\description{
Rename Columns based on Prefix
Used in func: createMoF
}
