% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{saveVar}
\alias{saveVar}
\title{saveVar}
\usage{
saveVar(data, fname)
}
\arguments{
\item{data}{Reactive dataframe}

\item{fname}{File name}
}
\description{
Save a dataframe in reactiveValues form into a json format. Written by Elie
}
