% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{Split.to.Tree}
\alias{Split.to.Tree}
\title{Split.to.Tree}
\usage{
Split.to.Tree(dtf)
}
\arguments{
\item{dtf}{dataframe with one or several dimensions. The last column of dtf must be the branch number}
}
\description{
Transform a dataframe into a list of list, as required for the first step to shape a Tree. Recursive function. Deprecated, use df2list
}
