% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{standardizeDF}
\alias{standardizeDF}
\title{Order a Data.frame (column-based)
Used in func: createLongFileEvent, createLongFileOcc}
\usage{
standardizeDF(df, finalCol)
}
\arguments{
\item{df}{a data.frame}

\item{finalCol}{a list of columns}
}
\value{
a data.frame
}
\description{
Order a Data.frame (column-based)
Used in func: createLongFileEvent, createLongFileOcc
}
