% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{Structure.to.Tree}
\alias{Structure.to.Tree}
\title{Structure.to.Tree}
\usage{
Structure.to.Tree(dtf)
}
\arguments{
\item{dtf}{list of (list of) dataframes. dtf must be the output of the function Split.to.Tree()}
}
\description{
Transform a list from Split.to.Tree() to the shape required by shinyTree(). Recursive function. Deprecated used df2list
}
