% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_func.R
\name{tidy_data}
\alias{tidy_data}
\title{Tidy User Data.Frame
Used in module: mod_Import}
\usage{
tidy_data(df, dfTidy)
}
\arguments{
\item{df}{a user data.frame to tidy}

\item{dfTidy}{a data.frame containing the information to tidy}
}
\value{
a tiy data.frame
}
\description{
Tidy User Data.Frame
Used in module: mod_Import
}
